feat: added process and addAlternative functions
The `susbstitute` function will be replaced by `addAlternative` and the
`transition` function will be updated to use `process`. This change will
allow for the code base to be split into two parts: those dependent on
starry-night and those not. Currently, `parse` and `addAlternative`
depend on starry-night, and `process` doesn't. Informally, any function
which receives code as text requires starry-night, but the goal is to
sequestor that functionality into its own functions so that they may be
preprocessed before runtime if `code-fns` is used on the web. The idea
is that a developer may use `babel-plugin-preval` to run the `parse`
function and its ilk and save the result to the built JavaScript output.
Then, the code may be displayed or transitioned at runtime without
downloading starry-night or its grammars. Ultimately, I would like to
build a macro to make this simpler.