feat: added tsdoc comments to most functions
The comments use a zero-width space in between the characters of `*/` to
allow me to use inline comments in my examples without ending the actual
tsdoc comment. I'll need to figure out how to remove this whitespace
later so that my syntax highlighting is correct in the documentation.
The tsdoc team is working on a supported way to escape such a sequence.
See [1]. I had to shut off the no-irregular-whitespace linting rule to
allow for this, but I think it's worth it for the time being. Some
editors may not show the zero-width space, but it's there.